fix(wintab): resolve pressure loss after first stroke
Root cause: WinTab context was being disabled by WebView/other apps,
causing pkNormalPressure to return 0 after the first stroke.

Solution: Periodically call WTEnable and WTOverlap (every 100 polling
loops) to keep the context active and at the top of the overlap order.

Changes:
- Add WTEnable/WTOverlap API bindings and periodic re-enable in polling loop
- Add Tauri environment detection to prevent invoke errors on startup
- Clear currentPoint on ProximityLeave for fresh stroke data
- Simplify PointerDown to use PointerEvent pressure (WinTab syncs in Move)
- Clean up debug logging, keep only essential diagnostics

+++ src-tauri/src/commands.rs
@@ -210,7 +210,7 @@ pub fn init_tablet(
             match window.window_handle() {
                 Ok(handle) => match handle.as_raw() {
                     RawWindowHandle::Win32(win32_handle) => {
-                        let hwnd = win32_handle.hwnd.get() as isize;
+                        let hwnd = win32_handle.hwnd.get();
                         tracing::info!("[Tablet] Got main window HWND: {}", hwnd);
                         Some(hwnd)
                     }
